Move Lower::HandleCall out to its own file.
