+++ toolchain/lower/BUILD
@@ -47,6 +47,7 @@ cc_library(
         "//toolchain/base:kind_switch",
         "//toolchain/sem_ir:entry_point",
         "//toolchain/sem_ir:file",
+        "//toolchain/sem_ir:ids",
         "//toolchain/sem_ir:inst",
         "//toolchain/sem_ir:inst_kind",
         "//toolchain/sem_ir:inst_namer",
